pyproject.toml
[project]
name = "tg-fastapi"
version = "0.1.0"
description = ""
authors = [
    { name = "Sweet_Sempai", email = "chcolatemilk00@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "tortoise-orm[asyncpg] (>=0.25.1,<0.26.0)",
    "isort (>=6.0.1,<7.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pylint (>=3.3.7,<4.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "aerich[toml] (>=0.9.0,<0.10.0)",
    "python-telegram-bot (>=22.1,<23.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
]

[build-system]
requires = [
    "poetry-core>=2.0.0,<3.0.0",
]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = [
    "py313",
]
include = "\\.pyi?$"
exclude = "/(\n    \\.git\n  | \\.venv\n  | build\n  | dist\n  | migrations\n  | __pycache__\n)/\n"

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = [
    "app",
    "bot",
    "db",
]
known_third_party = [
    "fastapi",
    "pydantic",
    "uvicorn",
    "python_telegram_bot",
    "telegram",
    "telegram.ext",
    "fastapi",
    "tortoise"
]
skip = [
    ".venv",
    "migrations",
    "__init__.py",
]

[tool.aerich]
tortoise_orm = "configs.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
